Fix for changing default class. Client crashed server on player spawn due to static fields in PERepresantive which caused TW code to find 2 constructors instead of one (one, correct, of type public and instanced and one of type private and static)
